Create Simple_detector_for_bivariate_gaussian_intensity_distribution
This code determines a bivariate Gaussian distriubtion, based on a time of measurement of an intensity rate (which is used to find expected intensity counts). A grid region is chosen in which the total intensity is found. This will change depending on the mean in x and y (which define the central peak value of the curve). A Gaussian process algorithm and neural network will then be used on this to maximise the area within the detector as the SNR value is lower (smaller times of measurement of the intensity).